unify the use of "unsigned int" use size_t

+++ src/align/trailing_comments.h
@@ -16,7 +16,7 @@
 class Chunk;
 class ChunkStack;
 
-enum class comment_align_e : unsigned int
+enum class comment_align_e : size_t
 {
    REGULAR,
    BRACE,
